FINERACT-2779: carry loan ids as Long through the loan tests
Follows the base helpers to Long: the loan-product locals lose their Integer declarations
and nineteen call sites lose the clientId.intValue() downcast.

Renaming one method is not cosmetic. ClientLoanChargeExternalIntegrationTest declared a
private applyForLoanApplication(Long, Long, String) whose third argument is a principal,
while the base method of the same shape takes an external id. On Integer ids the signatures
differed and the collision was invisible; it is now applyForLoanWithPrincipal.

Also drops a redundant String cast on getResourceExternalId, which the generated model has
returned as String for a while.

+++ integration-tests/src/test/java/org/apache/fineract/integrationtests/LoanTransactionInterestPaymentWaiverTest.java
@@ -762,7 +762,7 @@ public class LoanTransactionInterestPaymentWaiverTest extends FeignLoanTestBase 
         final Long clientCollateralId = collateralHelper.createClientCollateral(clientId, collateralId).getResourceId();
         assertNotNull(clientCollateralId);
 
-        final Integer productId = getLoanProductId(loanProductRequest);
+        final Long productId = getLoanProductId(loanProductRequest);
 
         final Long createActiveClientRequestId = 4730L;
         final Long applyLoanRequestId = createActiveClientRequestId + 1;
